refactor: calculate step methods in own module
